+++ mayan/apps/document_states/managers.py
@@ -4,4 +4,5 @@ from django.db import models
 class WorkflowManager(models.Manager):
     def launch_for(self, document):
         for workflow in document.document_type.workflows.all():
-            workflow.launch_for(document=document)
+            if workflow.auto_launch:
+                workflow.launch_for(document=document)