+++ mayan/apps/document_states/tests/mixins.py
@@ -170,7 +170,7 @@ class WorkflowStateViewTestMixin:
 class WorkflowTestMixin:
     def _create_test_workflow(self, add_document_type=False):
         self.test_workflow = Workflow.objects.create(
-            label=TEST_WORKFLOW_LABEL,
+            auto_launch=True, label=TEST_WORKFLOW_LABEL,
             internal_name=TEST_WORKFLOW_INTERNAL_NAME
         )
         self.test_workflow_runtime_proxy = WorkflowRuntimeProxy.objects.get(